+++ src/services/Products.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace craft\digitalproducts\services;
+
+use Craft;
+use craft\digitalproducts\elements\Product;
+use craft\events\SiteEvent;
+use craft\queue\jobs\ResaveElements;
+use yii\base\Component;
+
+/**
+ * Product service
+ *
+ * @author    Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @copyright Copyright (c) 2015, Pixel & Tonic, Inc.
+ */
+class Products extends Component
+{
+    // Public methods
+    // =========================================================================
+
+    /**
+     * Handle a Site being saved.
+     *
+     * @param SiteEvent $event
+     */
+    public function afterSaveSiteHandler(SiteEvent $event)
+    {
+        $queue = Craft::$app->getQueue();
+        $siteId = $event->oldPrimarySiteId;
+        $elementTypes = [
+            Product::class,
+        ];
+
+        foreach ($elementTypes as $elementType) {
+            $queue->push(new ResaveElements([
+                'elementType' => $elementType,
+                'criteria' => [
+                    'siteId' => $siteId,
+                    'status' => null,
+                    'enabledForSite' => false
+                ]
+            ]));
+        }
+    }
+}